Detect commands in issues and pull requests body
Before, it was necessary to add an extra comment with the `/vote` command to create a vote, as we were only processing comments events (in both issues and pull requests).

This PR adds the ability to also handle events for opened issues and pull requests, processing the corresponding entity body to detect commands. So it's possible now to launch a vote directly on the issue or pull request body without requiring an extra comment.

Closes #5

Signed-off-by: Sergio Castaño Arteaga <tegioz@icloud.com>